+++ components/SocialMediaPicker.tsx
@@ -1,5 +1,0 @@
-import { Pressable } from "react-native";
-
-export default function SocialMediaPicker() {
-  return <Pressable onPress={() => {}}></Pressable>;
-}